OOPification
Tested on the Nova server.
