+++ includes/watson.php
@@ -1,42 +1,6 @@
+<h2>Watson Upload</h2>
 <?php
 //Adapted from http://stackoverflow.com/questions/11448307/importing-csv-data-using-php-mysql
-
-//require_once('classes/connecti.php');
-
-
-//Get an array of all the countries.
-function getIdCountries(){
-	$connection = Database::getConnection();
-	
-	$query = "SELECT idcountry, country FROM country";
-	$result_obj="";
-	$items=array();
-	$result_obj=$connection->query($query);
-	try{
-		while($result = $result_obj->fetch_array(MYSQLI_ASSOC)){
-			$items[$result['idcountry']]=$result['country'];
-		}
-		return($items);
-	}
-	catch(Exception $e){
-		return false;
-	}	
-}
-
-//Empty the watson and watson_country tables. 
-function clearwatson(){
-	$connection = Database::getConnection();
-	//Empty watson table
-	$query="TRUNCATE TABLE watson";
-	if (!$connection->query($query)){
-		echo "Error :" .$query . "<br>" . $connection->error;
-	}
-	//Empty watson_country table
-	$query="TRUNCATE TABLE watson_country";
-	if (!$connection->query($query)){
-		echo "Error :" .$query . "<br>" . $connection->error;
-	}
-}
 
 //Upload File
 if (isset($_POST['submit'])) {
@@ -47,10 +11,12 @@ if (isset($_POST['submit'])) {
 	}
 
 	//Clear old watson data
-	clearwatson();
+	$subawards=new Subawards();
+	$subawards->clearwatson();
 	
 	//Get an array of all the countries
-	$allcountries=getIdCountries();
+	$country=new Country();
+	$allcountries=$country->getIdCountries();
 	//print_r($allcountries);
 	
 	$connection=Database::getConnection();
@@ -63,8 +29,6 @@ if (isset($_POST['submit'])) {
 		$budget=filter_var($budget, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
 		if ($i>1){	
 			$countriesarray=explode(',',$countries);
-			//print_r($countriesarray);
-			//echo "<br>";
 			$title=$connection->real_escape_string($title);
 			if ($shortawardnumber>1000 && $shortawardnumber<99999){
 				$query="INSERT INTO watson ";
@@ -89,19 +53,12 @@ if (isset($_POST['submit'])) {
 				//Add the countries to watson
 				foreach ($countriesarray as $country){
 					$country=trim($country);
-					//echo $country;
 					if ($country<>""){
 						$countryconnection=Database::getConnection();
 						$idcountry="";
 						//first need to get the idcountry for the country
 						$idcountry = array_search($country,$allcountries);
-						//echo $idcountry."<br>";
-						
-						//echo "hi".$idcountry."<br>";
-						//add it to the watson_country table.
 						$countryquery="INSERT INTO watson_country VALUES (DEFAULT, ".$idwatson.", ".$idcountry.")";
-						//echo $countryquery."<br>";
-						
 						if (!$countryconnection->query($countryquery)){
 							echo "Error :" .$countryquery . "<br>" . $countryconnection->error."<br>".$country;
 						}
